Moved login and register from usercontoller and userservice to authcontroller and authservice
