+++ src/main/java/com/roshan/filedigitizationsystem/controller/UserController.java
@@ -1,49 +1,21 @@
-package com.roshan.filedigitizationsystem.contoller;
+package com.roshan.filedigitizationsystem.controller;
 
-import com.roshan.filedigitizationsystem.dto.request.UserRequestDto;
-import com.roshan.filedigitizationsystem.dto.request.UserLoginDto;
 import com.roshan.filedigitizationsystem.dto.request.UserUpdateDto;
 import com.roshan.filedigitizationsystem.dto.response.UserResponseDto;
 import com.roshan.filedigitizationsystem.enums.UserRole;
 import com.roshan.filedigitizationsystem.exception.NoSuchUserExistsException;
-import com.roshan.filedigitizationsystem.exception.UserAlreadyExistsException;
 import com.roshan.filedigitizationsystem.service.UserServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("api/v1/users")
 public class UserController {
-    private final UserServiceImp userServiceImp;
-
     @Autowired
-    public UserController(UserServiceImp userServiceImp) {
-        this.userServiceImp = userServiceImp;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@RequestBody UserLoginDto userLoginDto) {
-        try {
-            UserResponseDto response = userServiceImp.login(userLoginDto);
-            return ResponseEntity.status(HttpStatus.OK).body(response);
-        } catch (BadCredentialsException ex) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ex.getMessage());
-        }
-
-    }
-
-    @PostMapping("/register")
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN','ROLE_MODERATOR')")
-    public ResponseEntity<?> register(@RequestBody UserRequestDto userRequestDto) {
-        userServiceImp.addUserInfo(userRequestDto);
-        return ResponseEntity.status(HttpStatus.CREATED).body("User has been registered");
-
-    }
-
+    private  UserServiceImp userServiceImp;
 
     @GetMapping
     @PreAuthorize("hasAnyRole('ROLE_ADMIN','ROLE_MODERATOR')")
